Move utility functions into specific file
